Fix rename operation
If the destination does not exist, Files.move throws an exception so it
is not fully equivalent to the File.renameTo method that was used before
which returns false in that case.

+++ java/org/apache/catalina/startup/HostConfig.java
@@ -1297,7 +1297,9 @@ public class HostConfig implements LifecycleListener {
             migration.execute();
 
             // Use rename
-            Files.move(destination.toPath(), tempOld.toPath());
+            if (destination.exists()) {
+                Files.move(destination.toPath(), tempOld.toPath());
+            }
             Files.move(tempNew.toPath(), destination.toPath());
             ExpandWar.delete(tempOld);
